Changed from a hard-coded c_double back to wp

+++ src/src_for/dnamiF.for
@@ -1,14 +1,14 @@
 subroutine time_march(param_int, param_float,data_float)&
 bind(c,name="time_march_fortran")
-  use iso_c_binding, only: c_double, c_int
+  use iso_c_binding, only: c_float, c_double, c_int
   implicit none  
 
 #include "dtypes.h"
 #include "param_fort.h" 
 
   integer(c_int),  intent(in)    ::   param_int(*)
-  real(c_double), intent(in)    :: param_float(*)
-  real(c_double), intent(inout) ::  data_float(*)
+  real(wp), intent(in)    :: param_float(*)
+  real(wp), intent(inout) ::  data_float(*)
 
   ! LOCAL VARIABLES
 
@@ -77,15 +77,15 @@ bind(c,name="time_march_fortran")
 
   subroutine stored(param_int,param_float,data_float,type_st)&
 bind(c,name="stored_fortran")
-  use iso_c_binding, only: c_double, c_int
+  use iso_c_binding, only: c_float, c_double, c_int
   implicit none  
 
 #include "dtypes.h"
 #include "param_fort.h" 
 
   integer,  intent(in)    ::   param_int(*)
-  real(c_double), intent(in)    :: param_float(*)
-  real(c_double), intent(inout) ::  data_float(*)
+  real(wp), intent(in)    :: param_float(*)
+  real(wp), intent(inout) ::  data_float(*)
   integer                 ::   type_st
  
  !f2py integer optional, intent(in)    :: type_st = 0
@@ -160,6 +160,7 @@ bind(c,name="stored_fortran")
                         qface_i , qface_j ,qface_k,&
                         qedge_ij, qedge_jk,qedge_ik)
 
+  use iso_c_binding, only: c_float, c_double, c_int
   implicit none
 
 #include "dtypes.h"
@@ -297,6 +298,7 @@ enddo ! END cache blocking k
                       qface_i , qface_j ,qface_k,&
                       qedge_ij, qedge_jk,qedge_ik)
 
+  use iso_c_binding, only: c_float, c_double, c_int
   implicit none
 
 #include "dtypes.h"
@@ -458,7 +460,7 @@ enddo ! END cache blocking k
 
 subroutine filter(dir,param_int,param_float,data_float)&
 bind(c,name="filter_fortran")
-  use iso_c_binding, only: c_double, c_int
+  use iso_c_binding, only: c_float, c_double, c_int
 implicit none  
 
 #include "dtypes.h"
@@ -466,8 +468,8 @@ implicit none
 
 integer(c_int), intent(in)   :: dir
 integer(c_int), intent(in)    :: param_int(*)
-real(c_double), intent(in)    :: param_float(*)
-real(c_double), intent(inout) ::  data_float(*)
+real(wp), intent(in)    :: param_float(*)
+real(wp), intent(inout) ::  data_float(*)
 
 !f2py  intent(in)   :: param_int,param_float,dir
 !f2py intent(inout) :: data_float    
@@ -515,6 +517,7 @@ real(c_double), intent(inout) ::  data_float(*)
 
 subroutine flt_x(param_float,hlo,neq,nx,ny,nz,sizeblck,nbc,bc,q,q2)
 
+use iso_c_binding, only: c_float, c_double, c_int
 implicit none  
 
 #include "dtypes.h"
@@ -635,6 +638,7 @@ idloop(6) = nz
 
 subroutine flt_y(param_float,hlo,neq,nx,ny,nz,sizeblck,nbc,bc,q,q2)
 
+use iso_c_binding, only: c_float, c_double, c_int
 implicit none  
 
 #include "dtypes.h"
@@ -759,6 +763,7 @@ idloop(6) = nz
 
  subroutine flt_z(param_float,hlo,neq,nx,ny,nz,sizeblck,nbc,bc,q,q2)
 
+use iso_c_binding, only: c_float, c_double, c_int
 implicit none  
 
 #include "dtypes.h"
@@ -882,6 +887,7 @@ idloop(6) = nz
 
 subroutine bc(dir,param_int,param_float,data_float)!param_float,hlo,nrk,neq,indvars,nx,ny,nz,q,q1,q2,rhs)
 
+use iso_c_binding, only: c_float, c_double, c_int
 implicit none  
 
 #include "dtypes.h"
@@ -960,6 +966,7 @@ integer, intent(in)   :: dir
 
 subroutine periodic_x(param_float,hlo,neq,nx,ny,nz,sizeblck,q,q2)
 
+use iso_c_binding, only: c_float, c_double, c_int
 implicit none  
 
 #include "dtypes.h"
@@ -1027,6 +1034,7 @@ enddo ! END cache blocking k
 
 subroutine periodic_y(param_float,hlo,neq,nx,ny,nz,sizeblck,q,q2)
 
+use iso_c_binding, only: c_float, c_double, c_int
 implicit none  
 
 #include "dtypes.h"
@@ -1096,6 +1104,7 @@ enddo ! END cache blocking k
 
  subroutine periodic_z(param_float,hlo,neq,nx,ny,nz,sizeblck,q,q2)
 
+use iso_c_binding, only: c_float, c_double, c_int
 implicit none  
 
 #include "dtypes.h"
@@ -1165,13 +1174,13 @@ idloop(1) = bi
 
 subroutine pack(buf,f,ibeg,iend,jbeg,jend,kbeg,kend,sizex,sizey,sizez,sizenv)&
 bind(c,name="pack_fortran")
-  use iso_c_binding, only: c_double, c_int
+  use iso_c_binding, only: c_float, c_double, c_int
 implicit none
 
 #include "dtypes.h"
 
-real(c_double), intent(out):: buf(*)
-real(c_double), intent(in) :: f(*)
+real(wp), intent(out):: buf(*)
+real(wp), intent(in) :: f(*)
 integer(c_int),intent(in)      :: ibeg,iend,jbeg,jend,kbeg,kend,sizex,sizey,sizez,sizenv
 
 integer :: sx,sy,sz,i,j,k,n,indbuf,indf
@@ -1204,13 +1213,13 @@ end subroutine pack
 
 subroutine unpack(buf,f,ibeg,iend,jbeg,jend,kbeg,kend,sizex,sizey,sizez,sizenv)&
 bind(c,name="unpack_fortran")
-  use iso_c_binding, only: c_double, c_int
+  use iso_c_binding, only: c_float, c_double, c_int
 implicit none
 
 #include "dtypes.h"
 
-real(c_double), intent(in) :: buf(*)
-real(c_double), intent(out) :: f(*)
+real(wp), intent(in) :: buf(*)
+real(wp), intent(out) :: f(*)
 integer(c_int),intent(in)      :: ibeg,iend,jbeg,jend,kbeg,kend,sizex,sizey,sizez,sizenv
 
 integer :: sx,sy,sz,i,j,k,n,indbuf,indf
@@ -1242,15 +1251,15 @@ end subroutine unpack
 
  subroutine init(param_int,param_float,data_float)&
 bind(c,name="init_fortran")
-  use iso_c_binding, only: c_double, c_int
+  use iso_c_binding, only: c_float, c_double, c_int
   implicit none  
 
 #include "dtypes.h"
 #include "param_fort.h" 
 
   integer(c_int),  intent(in)    :: param_int(*)
-  real(c_double), intent(in)    :: param_float(*)
-  real(c_double), intent(inout) ::  data_float(*)
+  real(wp), intent(in)    :: param_float(*)
+  real(wp), intent(inout) ::  data_float(*)
 
   call init_numa(param_float                                            ,&
                  param_int(iadrHLO), param_int(iadrNRK)                 ,&
@@ -1270,6 +1279,7 @@ bind(c,name="init_fortran")
 
 subroutine init_numa(param_float,hlo,nrk,neq,neqst,ind,nx,ny,nz,sizeblck,q,q1,q2,rhs,qst)
 
+  use iso_c_binding, only: c_float, c_double, c_int
   implicit none
 
 #include "dtypes.h"
